auth_service/requirements.txt
alembic==1.8.0
async-timeout==4.0.2
click==8.1.3
Deprecated==1.2.13
Flask==2.1.2
Flask-JWT-Extended==4.4.1
Flask-Login==0.6.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.16.0
marshmallow-sqlalchemy==0.28.0
packaging==21.3
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
python-slugify==6.1.2
redis==4.3.3
six==1.16.0
SQLAlchemy==1.4.37
text-unidecode==1.3
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
